// Api/models/team-member.ts
/* tslint:disable */
/* eslint-disable */
/**
 * hero-circle-api
 * The Hero Circle API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface TeamMember
 */
export interface TeamMember {

    /**
     * @type {string}
     * @memberof TeamMember
     */
    id: string;

    /**
     * @type {string}
     * @memberof TeamMember
     */
    name: string;

    /**
     * @type {string}
     * @memberof TeamMember
     */
    position: string;

    /**
     * @type {string}
     * @memberof TeamMember
     */
    image?: string;

    /**
     * @type {string}
     * @memberof TeamMember
     */
    country: string;

    /**
     * @type {string}
     * @memberof TeamMember
     */
    description: string;
}